Add Room behaviors, fix camera jerk on level start (#150)
* Add Room behaviors, fix camera jerk on level start

Rooms no longer spawn enemies by default.

Now, rooms provide the interface for doors
and adjacent rooms. Room behaviors define
what should happen within a room.

BossRoomBehavior:
Boss is now spawned under BossRoomBehavior
instead of in Level.

SpawnRoomBehavior:
The player spawn room will not spawn enemies,
and will automatically open its doors and
activate adjacent rooms.

EnemyRoomBehavior:
This was the default in all Rooms.
Rooms containing EnemyRoomBehavior will have
a Room > EnemyRoomBehavior > EnemySpawner node
structure. The EnemySpawner listens for the
Room's activation via the Room."activated"
signal. When the room is active, the spawner
will SpawnEnemies().

Other tweaks/bug fixes
- Enemies no longer become whiter with less health
- Remove camera jerk when starting game

* Remove debug messages, improve docs formatting

* Refactor SpawnRoomBehavior to StartingRoomBehavior
