+++ Coma Dash.csproj
@@ -73,7 +73,10 @@
     <Compile Include="Scripts\Generator\RegionLabel.cs" />
     <Compile Include="Scripts\Generator\RegionParseInfo.cs" />
     <Compile Include="Scripts\Generator\RoomGenerator.cs" />
-    <Compile Include="Scripts\Generator\Room.cs" />
+    <Compile Include="Scripts\Generator\Room\Room.cs" />
+    <Compile Include="Scripts\Generator\Room\RoomBehavior\BossRoomBehavior.cs" />
+    <Compile Include="Scripts\Generator\Room\RoomBehavior\EnemyRoomBehavior.cs" />
+    <Compile Include="Scripts\Generator\Room\RoomBehavior\StartingRoomBehavior.cs" />
     <Compile Include="Scripts\Generator\Wall.cs" />
     <Compile Include="Scripts\GUI.cs" />
     <Compile Include="Scripts\HealthEntity.cs" />
