+++ Scripts/Enemy/Enemy.cs
@@ -33,8 +33,7 @@ public class Enemy : HealthEntity
     public override void _Process(float delta)
     {
         Color baseColor = Colors.Red;
-        Color healthColor = baseColor.LinearInterpolate(Colors.White, 1 - Health / 100);
-        material.AlbedoColor = healthColor;
+        material.AlbedoColor = baseColor;
         UpdateStatusBars(delta);
     }
