+++ Scripts/Generator/Room/RoomBehavior/EnemyRoomBehavior.cs
@@ -1,0 +1,14 @@
+using Godot;
+
+public class EnemyRoomBehavior : Node
+{
+    private EnemySpawner enemySpawner;
+
+    public override void _Ready()
+    {
+        Room room = GetParent<Room>();
+        enemySpawner = new EnemySpawner();
+        AddChild(enemySpawner);
+        room.Connect("activated", enemySpawner, "OnRoomActivated");
+    }
+}
