+++ Scripts/Generator/Room/RoomBehavior/StartingRoomBehavior.cs
@@ -1,0 +1,12 @@
+using Godot;
+
+public class StartingRoomBehavior : Node
+{
+    public override void _Ready()
+    {
+        Room room = GetParent<Room>();
+        room.OpenAllConnectedDoors();
+        foreach (Room neighbor in room.GetConnectedRooms())
+            neighbor.Activate();
+    }
+}
